packaged request and response dto to avoid conflict and enhace readability
