+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/RequestService.java
@@ -1,6 +1,6 @@
 package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
-import org.orange.oie.internship2025.assetmanagementsystem.dto.RequestDTO;
-import org.orange.oie.internship2025.assetmanagementsystem.dto.ResponseDTO;
+import org.orange.oie.internship2025.assetmanagementsystem.dto.requestAsset.RequestDTO;
+import org.orange.oie.internship2025.assetmanagementsystem.dto.requestAsset.ResponseDTO;
 
 public interface RequestService {
